+++ Source/SwiftLintFramework/Extensions/Configuration+LintableFiles.swift
@@ -21,14 +21,12 @@ extension Configuration {
             return [path]
         }
         let pathsForPath = included.isEmpty ? fileManager.filesToLint(inPath: path, rootDirectory: nil) : []
-        let excludedPaths = excluded.flatMap {
-            fileManager.filesToLint(inPath: $0, rootDirectory: rootPath)
+        let excludedPaths = excluded.flatMap { excludedPath -> [String] in
+            return fileManager.filesToLint(inPath: excludedPath, rootDirectory: rootPath)
         }
         let includedPaths = included.flatMap {
             fileManager.filesToLint(inPath: $0, rootDirectory: rootPath)
         }
-        return (pathsForPath + includedPaths).filter {
-            !excludedPaths.contains($0)
-        }
+        return (pathsForPath + includedPaths).filter({ !excludedPaths.contains($0) })
     }
 }